refactor(plan-generator): TS AST 로직을 code-context-analyzer로 분리 (Plan C #C7)
- src/pipeline/phases/code-context-analyzer.ts 신규 (219줄)
  - collectContextualizationInfo 및 AST/regex 헬퍼 전량 이동
  - createSourceFile 공통 헬퍼로 AST 파일 생성 중복 제거
- src/pipeline/phases/plan-generator.ts: 772 → 468줄 (-304)
  - "import * as ts from typescript" 제거
  - readFileSync import도 정리
  - collectContextualizationInfo는 외부 호출 호환 위해 re-export

tsc 0, vitest 3342 pass (plan-generator.test.ts 30 pass).
