+++ src/pipeline/phases/plan-generator.ts
@@ -1,7 +1,7 @@
 import { resolve, join } from "path";
-import { readFileSync, existsSync, mkdirSync, writeFileSync } from "fs";
-import * as ts from "typescript";
+import { existsSync, mkdirSync, writeFileSync } from "fs";
 import { renderTemplate, loadTemplate, buildDynamicSection, extractDesignReferences, TemplateVariables } from "../../prompt/template-renderer.js";
+import { collectContextualizationInfo } from "./code-context-analyzer.js";
 import { detectCircularDependencies, validatePhaseDependencies } from "../execution/phase-scheduler.js";
 import { runClaude, extractJson } from "../../claude/claude-runner.js";
 import { configForTask, configForTaskWithMode } from "../../claude/model-router.js";
@@ -449,311 +449,7 @@ function validatePlan(plan: Plan): void {
   }
 }
 
-/**
- * Plan 생성 실패 시 관련 파일의 함수 시그니처와 import 관계를 수집한다.
- *
- * @param affectedFiles 영향받는 파일 경로 배열
- * @param cwd 작업 디렉토리
- * @returns 컨텍스트 정보
- */
-export function collectContextualizationInfo(affectedFiles: string[], cwd: string): ContextualizationInfo {
-  const functionSignatures: { [filePath: string]: string[] } = {};
-  const importRelations: { [filePath: string]: { imports: string[]; exports: string[] } } = {};
-  const typeDefinitions: { [filePath: string]: string[] } = {};
-
-  for (const filePath of affectedFiles) {
-    const fullPath = resolve(cwd, filePath);
-
-    // 파일이 존재하고 TypeScript/JavaScript 파일인 경우만 처리
-    if (!existsSync(fullPath) || !isTypeScriptOrJavaScriptFile(filePath)) {
-      continue;
-    }
-
-    try {
-      const content = readFileSync(fullPath, "utf-8");
-
-      functionSignatures[filePath] = extractFunctionSignatures(filePath, content);
-      importRelations[filePath] = extractImportRelations(filePath, content);
-      typeDefinitions[filePath] = extractTypeDefinitions(filePath, content);
-
-      logger.debug(`Collected context for ${filePath}: ${functionSignatures[filePath].length} functions, ${importRelations[filePath].imports.length} imports, ${typeDefinitions[filePath].length} types`);
-    } catch (error: unknown) {
-      logger.warn(`Failed to collect context for ${filePath}: ${getErrorMessage(error)}`);
-      // 실패한 파일은 빈 정보로 설정
-      functionSignatures[filePath] = [];
-      importRelations[filePath] = { imports: [], exports: [] };
-      typeDefinitions[filePath] = [];
-    }
-  }
-
-  return {
-    functionSignatures,
-    importRelations,
-    typeDefinitions,
-  };
-}
-
-/**
- * 파일이 TypeScript 또는 JavaScript 파일인지 확인
- */
-function isTypeScriptOrJavaScriptFile(filePath: string): boolean {
-  return /\.(ts|tsx|js|jsx)$/.test(filePath);
-}
-
-/**
- * TypeScript AST를 사용하여 함수 시그니처를 추출
- */
-function extractFunctionSignatures(filePath: string, content: string): string[] {
-  const signatures: string[] = [];
-
-  try {
-    const sourceFile = ts.createSourceFile(
-      filePath,
-      content,
-      ts.ScriptTarget.Latest,
-      true,
-      filePath.endsWith('.tsx') || filePath.endsWith('.jsx') ? ts.ScriptKind.TSX : ts.ScriptKind.TS
-    );
-
-    visitNodeForSignatures(sourceFile, signatures);
-  } catch (_error: unknown) {
-    // TypeScript 파싱 실패 시 정규식으로 폴백
-    return extractFunctionSignaturesRegex(content);
-  }
-
-  return signatures;
-}
-
-/**
- * AST 노드를 순회하여 함수 시그니처 수집
- */
-function visitNodeForSignatures(node: ts.Node, signatures: string[]): void {
-  if (ts.isFunctionDeclaration(node) && node.name) {
-    const signature = getFunctionSignature(node);
-    if (signature) {
-      signatures.push(signature);
-    }
-  } else if (ts.isMethodDeclaration(node) || ts.isMethodSignature(node)) {
-    const signature = getMethodSignature(node);
-    if (signature) {
-      signatures.push(signature);
-    }
-  } else if (ts.isArrowFunction(node) || ts.isFunctionExpression(node)) {
-    // 변수에 할당된 화살표 함수나 함수 표현식 처리
-    const parent = node.parent;
-    if (ts.isVariableDeclaration(parent) && parent.name && ts.isIdentifier(parent.name)) {
-      const signature = `${parent.name.text}: ${node.getText().substring(0, 100)}...`;
-      signatures.push(signature);
-    }
-  }
-
-  ts.forEachChild(node, (child) => visitNodeForSignatures(child, signatures));
-}
-
-/**
- * TypeScript AST에서 함수 시그니처 추출
- */
-function getFunctionSignature(node: ts.FunctionDeclaration): string | null {
-  if (!node.name) return null;
-
-  const name = node.name.text;
-  const params = node.parameters.map(param => {
-    const paramName = param.name.getText();
-    const paramType = param.type ? param.type.getText() : 'any';
-    return `${paramName}: ${paramType}`;
-  }).join(', ');
-
-  const returnType = node.type ? node.type.getText() : 'void';
-  const asyncKeyword = node.modifiers?.some(m => m.kind === ts.SyntaxKind.AsyncKeyword) ? 'async ' : '';
-  const exportKeyword = node.modifiers?.some(m => m.kind === ts.SyntaxKind.ExportKeyword) ? 'export ' : '';
-
-  return `${exportKeyword}${asyncKeyword}function ${name}(${params}): ${returnType}`;
-}
-
-/**
- * TypeScript AST에서 메소드 시그니처 추출
- */
-function getMethodSignature(node: ts.MethodDeclaration | ts.MethodSignature): string | null {
-  const name = node.name?.getText();
-  if (!name) return null;
-
-  const params = node.parameters.map(param => {
-    const paramName = param.name.getText();
-    const paramType = param.type ? param.type.getText() : 'any';
-    return `${paramName}: ${paramType}`;
-  }).join(', ');
-
-  const returnType = node.type ? node.type.getText() : 'void';
-  const asyncKeyword = node.modifiers?.some(m => m.kind === ts.SyntaxKind.AsyncKeyword) ? 'async ' : '';
-
-  return `${asyncKeyword}${name}(${params}): ${returnType}`;
-}
-
-/**
- * 정규식을 사용한 함수 시그니처 추출 (폴백)
- */
-function extractFunctionSignaturesRegex(content: string): string[] {
-  const signatures: string[] = [];
-
-  // function 선언
-  const functionRegex = /(?:export\s+)?(?:async\s+)?function\s+(\w+)\s*\([^)]*\)(?:\s*:\s*[^{]+)?/g;
-  let match;
-  while ((match = functionRegex.exec(content)) !== null) {
-    signatures.push(match[0]);
-  }
-
-  // 화살표 함수 (변수 할당)
-  const arrowFunctionRegex = /(?:export\s+)?const\s+(\w+)\s*[:=]\s*(?:async\s+)?\([^)]*\)\s*=>/g;
-  while ((match = arrowFunctionRegex.exec(content)) !== null) {
-    signatures.push(match[0] + '...');
-  }
-
-  return signatures;
-}
-
-/**
- * Import/Export 관계 추출
- */
-function extractImportRelations(filePath: string, content: string): { imports: string[]; exports: string[] } {
-  const imports: string[] = [];
-  const exports: string[] = [];
-
-  try {
-    const sourceFile = ts.createSourceFile(
-      filePath,
-      content,
-      ts.ScriptTarget.Latest,
-      true,
-      filePath.endsWith('.tsx') || filePath.endsWith('.jsx') ? ts.ScriptKind.TSX : ts.ScriptKind.TS
-    );
-
-    visitNodeForImports(sourceFile, imports, exports);
-  } catch (_error: unknown) {
-    // 파싱 실패 시 정규식으로 폴백
-    return extractImportRelationsRegex(content);
-  }
-
-  return { imports, exports };
-}
-
-/**
- * AST 노드를 순회하여 import/export 관계 수집
- */
-function visitNodeForImports(node: ts.Node, imports: string[], exports: string[]): void {
-  if (ts.isImportDeclaration(node)) {
-    const importText = node.getText();
-    imports.push(importText);
-  } else if (ts.isExportDeclaration(node) || ts.isExportAssignment(node)) {
-    const exportText = node.getText();
-    exports.push(exportText);
-  } else if (ts.canHaveModifiers(node) && ts.getModifiers(node)?.some((m: ts.Modifier) => m.kind === ts.SyntaxKind.ExportKeyword)) {
-    // export 키워드가 있는 선언들
-    const exportText = node.getText().substring(0, 100) + '...';
-    exports.push(exportText);
-  }
-
-  ts.forEachChild(node, (child) => visitNodeForImports(child, imports, exports));
-}
-
-/**
- * 정규식을 사용한 Import/Export 관계 추출 (폴백)
- */
-function extractImportRelationsRegex(content: string): { imports: string[]; exports: string[] } {
-  const imports: string[] = [];
-  const exports: string[] = [];
-
-  // import 문 추출
-  const importRegex = /import\s+.*?from\s+["'].*?["'];?/g;
-  let match;
-  while ((match = importRegex.exec(content)) !== null) {
-    imports.push(match[0]);
-  }
-
-  // export 문 추출
-  const exportRegex = /export\s+.*?(?:[;}]|$)/gm;
-  while ((match = exportRegex.exec(content)) !== null) {
-    exports.push(match[0]);
-  }
-
-  return { imports, exports };
-}
-
-/**
- * 타입 정의 추출
- */
-function extractTypeDefinitions(filePath: string, content: string): string[] {
-  const typeDefinitions: string[] = [];
-
-  try {
-    const sourceFile = ts.createSourceFile(
-      filePath,
-      content,
-      ts.ScriptTarget.Latest,
-      true,
-      filePath.endsWith('.tsx') || filePath.endsWith('.jsx') ? ts.ScriptKind.TSX : ts.ScriptKind.TS
-    );
-
-    visitNodeForTypes(sourceFile, typeDefinitions);
-  } catch (_error: unknown) {
-    // 파싱 실패 시 정규식으로 폴백
-    return extractTypeDefinitionsRegex(content);
-  }
-
-  return typeDefinitions;
-}
-
-/**
- * AST 노드를 순회하여 타입 정의 수집
- */
-function visitNodeForTypes(node: ts.Node, typeDefinitions: string[]): void {
-  if (ts.isTypeAliasDeclaration(node)) {
-    typeDefinitions.push(node.getText());
-  } else if (ts.isInterfaceDeclaration(node)) {
-    typeDefinitions.push(node.getText());
-  } else if (ts.isEnumDeclaration(node)) {
-    typeDefinitions.push(node.getText());
-  } else if (ts.isClassDeclaration(node)) {
-    const className = node.name?.text || 'Unknown';
-    const classSignature = `class ${className}${node.heritageClauses ? ' extends/implements...' : ''}`;
-    typeDefinitions.push(classSignature);
-  }
-
-  ts.forEachChild(node, (child) => visitNodeForTypes(child, typeDefinitions));
-}
-
-/**
- * 정규식을 사용한 타입 정의 추출 (폴백)
- */
-function extractTypeDefinitionsRegex(content: string): string[] {
-  const typeDefinitions: string[] = [];
-
-  // interface 정의
-  const interfaceRegex = /(?:export\s+)?interface\s+\w+\s*(?:<[^>]*>)?\s*(?:extends\s+[^{]*)?\s*\{[^}]*\}/gs;
-  let match;
-  while ((match = interfaceRegex.exec(content)) !== null) {
-    typeDefinitions.push(match[0]);
-  }
-
-  // type 정의
-  const typeRegex = /(?:export\s+)?type\s+\w+(?:<[^>]*>)?\s*=\s*[^;]+;?/g;
-  while ((match = typeRegex.exec(content)) !== null) {
-    typeDefinitions.push(match[0]);
-  }
-
-  // enum 정의
-  const enumRegex = /(?:export\s+)?enum\s+\w+\s*\{[^}]*\}/gs;
-  while ((match = enumRegex.exec(content)) !== null) {
-    typeDefinitions.push(match[0]);
-  }
-
-  // class 정의
-  const classRegex = /(?:export\s+)?class\s+\w+(?:\s+extends\s+\w+)?(?:\s+implements\s+[^{]*)?\s*\{/g;
-  while ((match = classRegex.exec(content)) !== null) {
-    typeDefinitions.push(match[0] + '...}');
-  }
-
-  return typeDefinitions;
-}
+export { collectContextualizationInfo };
 
 /**
  * retry 템플릿에서 retry 특화 섹션만 추출합니다.
